simplify to just .list()
